Remove SecUserDetail, reformat User
